requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
oauthlib==3.2.2
protobuf==4.21.12
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
tweepy==4.14.0
urllib3==2.0.4
Werkzeug==2.3.6
